baselineGL-V1, Baseline Local update

+++ other_tracker/Baseline/lib/models/stark/backbone.py
@@ -80,7 +80,7 @@ class BackboneBase(nn.Module):
         self.num_channels = num_channels
 
     def forward(self, tensor_list: NestedTensor):
-        xs = self.body(tensor_list.tensors)
+        xs = self.body(tensor_list.tensors.to(torch.float32))
         out: Dict[str, NestedTensor] = {}
         for name, x in xs.items():
             m = tensor_list.mask
